requirements.txt
streamlit==1.27.2
google-generativeai==0.3.1
openai==1.4.0
